add tests for replay
